+++ src/plans/objective.jl
@@ -15,7 +15,7 @@ All these elements should usually be implemented as functions
 `(M, p) -> ...`, or `(M, X, p) -> ...` that is
 
 * the first argument of these functions should be the manifold `M` they are defined on
-* the argument `X` is present, if the computation is performed inplace of `X` (see [`InplaceEvaluation`](@ref))
+* the argument `X` is present, if the computation is performed in-place of `X` (see [`InplaceEvaluation`](@ref))
 * the argument `p` is the place the function (``f`` or one of its elements) is evaluated __at__.
 
 the type `T` indicates the global [`AbstractEvaluationType`](@ref).